+++ controller/processor.go
@@ -3,10 +3,12 @@ package controller
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/tools/cache"
-	"k8s.io/client-go/util/workqueue"
+
+	"github.com/spotahome/kooper/log"
 )
 
 // processor knows how to process object keys.
@@ -14,72 +16,62 @@ type processor interface {
 	Process(ctx context.Context, key string) error
 }
 
+// processorFunc a helper to create processors.
+type processorFunc func(ctx context.Context, key string) error
+
+func (p processorFunc) Process(ctx context.Context, key string) error { return p(ctx, key) }
+
+// newIndexerProcessor returns a processor that processes a key that will get the kubernetes object
+// from a cache called indexer were the kubernetes watch updates have been indexed and stored
+// by the listerwatchers from the informers.
 func newIndexerProcessor(indexer cache.Indexer, handler Handler) processor {
-	return indexerProcessor{
-		indexer: indexer,
-		handler: handler,
-	}
-}
+	return processorFunc(func(ctx context.Context, key string) error {
+		// Get the object
+		obj, exists, err := indexer.GetByKey(key)
+		if err != nil {
+			return err
+		}
 
-// indexerProcessor processes a key that will get the kubernetes object from a cache
-// called indexer were the kubernetes watch updates have been indexed and stored by the
-// listerwatchers from the informers.
-type indexerProcessor struct {
-	indexer cache.Indexer
-	handler Handler
-}
+		// Handle the object.
+		if !exists {
+			return handler.Delete(ctx, key)
+		}
 
-func (i indexerProcessor) Process(ctx context.Context, key string) error {
-	// Get the object
-	obj, exists, err := i.indexer.GetByKey(key)
-	if err != nil {
-		return err
-	}
-
-	// handle the object.
-	if !exists { // Deleted resource from the cache.
-		return i.handler.Delete(ctx, key)
-	}
-
-	return i.handler.Add(ctx, obj.(runtime.Object))
+		return handler.Add(ctx, obj.(runtime.Object))
+	})
 }
 
 var errRequeued = fmt.Errorf("requeued after receiving error")
 
-// retryProcessor will delegate the processing of a key to the received processor,
-// in case the processing/handling of this key fails it will add the key
+// newRetryProcessor returns a processor that will delegate the processing of a key to the
+// received processor, in case the processing/handling of this key fails it will add the key
 // again to a queue if it has retrys pending.
 //
-// If the processing errored and has been retried, it will return a `errRequeued`
-// error.
-type retryProcessor struct {
-	name       string
-	maxRetries int
-	mrec       MetricsRecorder
-	queue      workqueue.RateLimitingInterface
-	next       processor
+// If the processing errored and has been retried, it will return a `errRequeued` error.
+func newRetryProcessor(name string, maxRetries int, queue blockingQueue, logger log.Logger, next processor) processor {
+	return processorFunc(func(ctx context.Context, key string) error {
+		err := next.Process(ctx, key)
+		if err != nil {
+			// Retry if possible.
+			requeueErr := queue.Requeue(ctx, key)
+			if requeueErr != nil {
+				return fmt.Errorf("could not retry: %s: %w", requeueErr, err)
+			}
+			logger.WithKV(log.KV{"object-key": key}).Warningf("item requeued due to processing error: %s", err)
+			return nil
+		}
+
+		return nil
+	})
 }
 
-func newRetryProcessor(name string, maxRetries int, mrec MetricsRecorder, queue workqueue.RateLimitingInterface, next processor) processor {
-	return retryProcessor{
-		name:       name,
-		maxRetries: maxRetries,
-		mrec:       mrec,
-		queue:      queue,
-		next:       next,
-	}
-}
+// newMetricsProcessor returns a processor that measures everything related with the processing logic.
+func newMetricsProcessor(name string, mrec MetricsRecorder, next processor) processor {
+	return processorFunc(func(ctx context.Context, key string) (err error) {
+		defer func(t0 time.Time) {
+			mrec.ObserveResourceProcessingDuration(ctx, name, err == nil, t0)
+		}(time.Now())
 
-func (r retryProcessor) Process(ctx context.Context, key string) error {
-	err := r.next.Process(ctx, key)
-
-	// If there was an error and we have retries pending then requeue.
-	if err != nil && r.queue.NumRequeues(key) < r.maxRetries {
-		r.queue.AddRateLimited(key)
-		r.mrec.IncResourceEventQueued(ctx, r.name, RequeueEvent)
-		return fmt.Errorf("%w: %s", errRequeued, err)
-	}
-
-	r.queue.Forget(key)
-	return err
+		return next.Process(ctx, key)
+	})
 }