+++ metrics/prometheus/prometheus.go
@@ -2,6 +2,7 @@ package prometheus
 
 import (
 	"context"
+	"strconv"
 	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
@@ -18,11 +19,12 @@ type Config struct {
 	// Registerer is a prometheus registerer, e.g: prometheus.Registry.
 	// By default will use Prometheus default registry.
 	Registerer prometheus.Registerer
-	// Buckets sets custom buckets for the duration/latency metrics. This should be used when
-	// the default buckets don't work. This could happen when the time to process an event is not on the
-	// range of 5ms-10s duration.
+	// InQueueBuckets sets custom buckets for the duration/latency items in queue metrics.
 	// Check https://godoc.org/github.com/prometheus/client_golang/prometheus#pkg-variables
-	Buckets []float64
+	InQueueBuckets []float64
+	// ProcessingBuckets sets custom buckets for the duration/latency processing metrics.
+	// Check https://godoc.org/github.com/prometheus/client_golang/prometheus#pkg-variables
+	ProcessingBuckets []float64
 }
 
 func (c *Config) defaults() {
@@ -30,17 +32,19 @@ func (c *Config) defaults() {
 		c.Registerer = prometheus.DefaultRegisterer
 	}
 
-	if c.Buckets == nil || len(c.Buckets) == 0 {
-		c.Buckets = prometheus.DefBuckets
+	if c.InQueueBuckets == nil || len(c.InQueueBuckets) == 0 {
+		c.InQueueBuckets = prometheus.DefBuckets
+	}
+
+	if c.ProcessingBuckets == nil || len(c.ProcessingBuckets) == 0 {
+		c.ProcessingBuckets = prometheus.DefBuckets
 	}
 }
 
 // Recorder implements the metrics recording in a prometheus registry.
 type Recorder struct {
-	// Metrics
-	queuedEvents           *prometheus.CounterVec
-	processedEvents        *prometheus.CounterVec
-	processedEventErrors   *prometheus.CounterVec
+	queuedEventsTotal      *prometheus.CounterVec
+	inQueueEventDuration   *prometheus.HistogramVec
 	processedEventDuration *prometheus.HistogramVec
 }
 
@@ -49,66 +53,55 @@ func New(cfg Config) *Recorder {
 	cfg.defaults()
 
 	r := &Recorder{
-		queuedEvents: prometheus.NewCounterVec(prometheus.CounterOpts{
+		queuedEventsTotal: prometheus.NewCounterVec(prometheus.CounterOpts{
 			Namespace: promNamespace,
 			Subsystem: promControllerSubsystem,
 			Name:      "queued_events_total",
 			Help:      "Total number of events queued.",
-		}, []string{"controller", "type"}),
+		}, []string{"controller", "requeue"}),
 
-		processedEvents: prometheus.NewCounterVec(prometheus.CounterOpts{
+		inQueueEventDuration: prometheus.NewHistogramVec(prometheus.HistogramOpts{
 			Namespace: promNamespace,
 			Subsystem: promControllerSubsystem,
-			Name:      "processed_events_total",
-			Help:      "Total number of successfuly processed events.",
-		}, []string{"controller", "type"}),
-
-		processedEventErrors: prometheus.NewCounterVec(prometheus.CounterOpts{
-			Namespace: promNamespace,
-			Subsystem: promControllerSubsystem,
-			Name:      "processed_event_errors_total",
-			Help:      "Total number of errors processing events.",
-		}, []string{"controller", "type"}),
+			Name:      "event_in_queue_duration_total",
+			Help:      "The duration of an event in the queue.",
+			Buckets:   cfg.InQueueBuckets,
+		}, []string{"controller"}),
 
 		processedEventDuration: prometheus.NewHistogramVec(prometheus.HistogramOpts{
 			Namespace: promNamespace,
 			Subsystem: promControllerSubsystem,
 			Name:      "processed_event_duration_seconds",
-			Help:      "The duration for a successful event to be processed.",
-			Buckets:   cfg.Buckets,
-		}, []string{"controller", "type"}),
+			Help:      "The duration for an event to be processed.",
+			Buckets:   cfg.ProcessingBuckets,
+		}, []string{"controller", "success"}),
 	}
 
 	// Register metrics.
 	cfg.Registerer.MustRegister(
-		r.queuedEvents,
-		r.processedEvents,
-		r.processedEventErrors,
+		r.queuedEventsTotal,
+		r.inQueueEventDuration,
 		r.processedEventDuration)
 
 	return r
 }
 
-// IncResourceEventQueued satisfies metrics.Recorder interface.
-func (r Recorder) IncResourceEventQueued(_ context.Context, controller string, eventType controller.EventType) {
-	r.queuedEvents.WithLabelValues(controller, string(eventType)).Inc()
+// IncResourceEventQueued satisfies controller.MetricsRecorder interface.
+func (r Recorder) IncResourceEventQueued(ctx context.Context, controller string, isRequeue bool) {
+	r.queuedEventsTotal.WithLabelValues(controller, strconv.FormatBool(isRequeue)).Inc()
 }
 
-// IncResourceEventProcessed satisfies metrics.Recorder interface.
-func (r Recorder) IncResourceEventProcessed(_ context.Context, controller string, eventType controller.EventType) {
-	r.processedEvents.WithLabelValues(controller, string(eventType)).Inc()
+// ObserveResourceInQueueDuration satisfies controller.MetricsRecorder interface.
+func (r Recorder) ObserveResourceInQueueDuration(ctx context.Context, controller string, queuedAt time.Time) {
+	r.inQueueEventDuration.WithLabelValues(controller).
+		Observe(time.Since(queuedAt).Seconds())
 }
 
-// IncResourceEventProcessedError satisfies metrics.Recorder interface.
-func (r Recorder) IncResourceEventProcessedError(_ context.Context, controller string, eventType controller.EventType) {
-	r.processedEventErrors.WithLabelValues(controller, string(eventType)).Inc()
+// ObserveResourceProcessingDuration satisfies controller.MetricsRecorder interface.
+func (r Recorder) ObserveResourceProcessingDuration(ctx context.Context, controller string, success bool, startProcessingAt time.Time) {
+	r.processedEventDuration.WithLabelValues(controller, strconv.FormatBool(success)).
+		Observe(time.Since(startProcessingAt).Seconds())
 }
 
-// ObserveDurationResourceEventProcessed satisfies metrics.Recorder interface.
-func (r Recorder) ObserveDurationResourceEventProcessed(_ context.Context, controller string, eventType controller.EventType, start time.Time) {
-	secs := time.Now().Sub(start).Seconds()
-	r.processedEventDuration.WithLabelValues(controller, string(eventType)).Observe(secs)
-}
-
-// Check we implement all the required metrics recorder interfaces.
+// Check interfaces implementation.
 var _ controller.MetricsRecorder = &Recorder{}